Moved log file parsing to a common header.
